requirements.txt
amqp==5.0.8
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==2.2
django-cors-headers==3.10.1
djangorestframework==3.12.4
kombu==5.2.2
prompt-toolkit==3.0.24
psycopg2==2.8.6
pycodestyle==2.8.0
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
